embedding_service/requirements.txt
# ===================================
# EMBEDDING SERVICE - CPU ONLY
# Verificado: Febrero 2026
# Python 3.11.9 requerido
# ===================================

# ----------------------------------
# PYTORCH CPU (Sin CUDA)
# ----------------------------------
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu
torchvision==0.25.0+cpu
torchaudio==2.10.0+cpu

# ----------------------------------
# VECTORES Y EMBEDDINGS
# ----------------------------------
# FAISS para busqueda vectorial (CPU)
faiss-cpu==1.13.2

# Sentence Transformers (generacion de embeddings)
sentence-transformers==5.2.2

# Dependencias de sentence-transformers
transformers>=4.45.0
huggingface-hub>=0.26.0
tokenizers>=0.20.0

# ----------------------------------
# PROCESAMIENTO NUMERICO
# ----------------------------------
numpy>=1.26.4,<2.3.0
scikit-learn>=1.6.1
scipy>=1.11.0

# ----------------------------------
# FRAMEWORK WEB
# ----------------------------------
fastapi==0.115.12
uvicorn[standard]==0.33.0
pydantic==2.10.6
pydantic-settings==2.7.1

# ----------------------------------
# gRPC
# ----------------------------------
grpcio==1.69.0
grpcio-tools==1.69.0

# ----------------------------------
# UTILIDADES
# ----------------------------------
python-multipart==0.0.20
typing-extensions>=4.12.0
pyyaml==6.0.2
python-json-logger==3.2.1
tqdm>=4.66.0
cachetools==7.0.1

# Monitoreo de CPU y RAM
psutil==6.1.1

# ----------------------------------
# TESTING
# ----------------------------------
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
pytest-cov==6.0.0

# ----------------------------------
# DESARROLLO (Opcional)
# ----------------------------------
black==25.1.0
flake8==7.1.1
mypy==1.14.1